+++ SystemSimulation/TextScreen.h
@@ -57,7 +57,7 @@ extern struct termios orig_termios;
 void reset_terminal_mode();
 void set_conio_terminal_mode();
 int kbhit();
-int getch();
+u8 getch();
 
 struct ColourChar
 {
